some first cleanup
