+++ src/framing.rs
@@ -177,7 +177,7 @@ impl MLSCiphertext {
             epoch: context.epoch,
             content_type: mls_plaintext.content_type,
             authenticated_data: mls_plaintext.authenticated_data.to_vec(),
-            sender_data_nonce: sender_data_nonce.as_slice(),
+            sender_data_nonce: sender_data_nonce.as_slice().to_vec(),
         };
         let mls_ciphertext_sender_data_aad_bytes =
             mls_ciphertext_sender_data_aad.encode_detached().unwrap(); // TODO: error handling
@@ -194,7 +194,7 @@ impl MLSCiphertext {
             epoch: context.epoch,
             content_type: mls_plaintext.content_type,
             authenticated_data: mls_plaintext.authenticated_data.to_vec(),
-            sender_data_nonce: sender_data_nonce.as_slice(),
+            sender_data_nonce: sender_data_nonce.as_slice().to_vec(),
             encrypted_sender_data: encrypted_sender_data.clone(),
         };
         let mls_ciphertext_content_aad_bytes =
@@ -212,7 +212,7 @@ impl MLSCiphertext {
             + mls_plaintext.content.encode_detached().unwrap().len()
             + mls_plaintext.signature.encode_detached().unwrap().len()
             + 2
-            + TAGBYTES
+            + TAG_BYTES
             + 4;
         let mut padding_length = (config.padding_block_size as usize)
             - (padding_offset % (config.padding_block_size as usize));
@@ -233,7 +233,7 @@ impl MLSCiphertext {
                     &epoch_secrets.handshake_secret,
                     "hs nonce",
                     &mls_plaintext.sender.encode_detached().unwrap(),
-                    NONCEBYTES,
+                    NONCE_BYTES,
                 );
                 let reuse_guard = sender_data.reuse_guard.encode_detached().unwrap();
                 for i in 0..4 {
@@ -245,7 +245,7 @@ impl MLSCiphertext {
                     &epoch_secrets.handshake_secret,
                     "hs key",
                     &mls_plaintext.sender.encode_detached().unwrap(),
-                    CHACHAKEYBYTES,
+                    CHACHA_KEY_BYTES,
                 );
                 let handshake_key = ciphersuite.new_aead_key(&handshake_key_input).unwrap();
                 (handshake_key, handshake_nonce)
@@ -264,7 +264,7 @@ impl MLSCiphertext {
             epoch: context.epoch,
             content_type: mls_plaintext.content_type,
             authenticated_data: mls_plaintext.authenticated_data.to_vec(),
-            sender_data_nonce: sender_data_nonce.as_slice(),
+            sender_data_nonce: sender_data_nonce.as_slice().to_vec(),
             encrypted_sender_data,
             ciphertext,
         }
@@ -292,7 +292,7 @@ impl MLSCiphertext {
             epoch: self.epoch,
             content_type: self.content_type,
             authenticated_data: self.authenticated_data.clone(),
-            sender_data_nonce: sender_data_nonce.as_slice(),
+            sender_data_nonce: sender_data_nonce.as_slice().to_vec(),
         };
         let mls_ciphertext_sender_data_aad_bytes =
             mls_ciphertext_sender_data_aad.encode_detached().unwrap();
@@ -313,7 +313,7 @@ impl MLSCiphertext {
             epoch: self.epoch,
             content_type: self.content_type,
             authenticated_data: self.authenticated_data.clone(),
-            sender_data_nonce: sender_data_nonce.as_slice(),
+            sender_data_nonce: sender_data_nonce.as_slice().to_vec(),
             encrypted_sender_data: self.encrypted_sender_data.clone(),
         };
         let mls_ciphertext_content_aad_bytes =
@@ -326,7 +326,7 @@ impl MLSCiphertext {
                     &epoch_secrets.handshake_secret,
                     "hs nonce",
                     &sender_data.sender.as_u32().encode_detached().unwrap(),
-                    NONCEBYTES,
+                    NONCE_BYTES,
                 );
                 let reuse_guard = sender_data.reuse_guard.encode_detached().unwrap();
                 for i in 0..4 {
@@ -338,7 +338,7 @@ impl MLSCiphertext {
                     &epoch_secrets.handshake_secret,
                     "hs key",
                     &sender_data.sender.as_u32().encode_detached().unwrap(),
-                    CHACHAKEYBYTES,
+                    CHACHA_KEY_BYTES,
                 );
                 let handshake_key = ciphersuite.new_aead_key(&handshake_key_input).unwrap();
                 (handshake_key, handshake_nonce)
@@ -806,7 +806,7 @@ impl From<MLSPlaintext> for MLSPlaintextCommitAuthData {
         };
         MLSPlaintextCommitAuthData {
             confirmation: confirmation.0,
-            signature: mls_plaintext.signature.as_slice(),
+            signature: mls_plaintext.signature.as_slice().to_vec(),
         }
     }
 }
@@ -829,7 +829,7 @@ impl Codec for MLSPlaintextCommitAuthData {
 
 #[test]
 fn codec() {
-    let ciphersuite = Ciphersuite::new(Name::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519);
+    let ciphersuite = Ciphersuite::new(CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519);
     let keypair = ciphersuite.new_signature_keypair();
     let sender = Sender {
         sender_type: SenderType::Member,
@@ -862,8 +862,13 @@ fn codec() {
 fn padding() {
     use crate::ciphersuite::*;
     use crate::utils::*;
+    use rand::RngCore;
 
-    let ciphersuite = Ciphersuite::new(Name::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519);
+    fn random_usize() -> usize {
+        rand::rngs::OsRng.next_u64() as usize
+    }
+
+    let ciphersuite = Ciphersuite::new(CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519);
     let alice_identity = Identity::new(ciphersuite, vec![1, 2, 3]);
 
     let config = GROUP_CONFIG_DEFAULT;
