+++ src/lib.rs
@@ -24,7 +24,6 @@ pub mod group;
 pub mod kp;
 pub mod messages;
 pub mod schedule;
-pub mod tree;
-pub mod treemath;
-pub mod utils;
+mod tree;
+mod utils;
 pub mod validator;
